requirements.txt
tabulate==0.8.9
hyperopt==0.2.5
tqdm~=4.65.0
geopy==2.0.0
dgl==0.6.1
fastdtw==0.3.4
networkx~=3.1
torchtext==0.6.0
nltk==3.2.4
tensorboardx==2.2
aioredis==1.3.1
aiohttp==3.7.4
idna==2.10
torchdiffeq==0.2.3
pandas~=1.4.4
numpy~=1.24.3
scikit-learn~=1.3.0
scipy~=1.10.1
torchvision~=0.16.1